requirements.txt
streamlit==1.35.0
Pillow>=10.0